+++ envoy-control-core/src/test/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/SnapshotUpdaterTest.kt
@@ -41,6 +41,7 @@ import pl.allegro.tech.servicemesh.envoycontrol.services.MultiClusterState
 import pl.allegro.tech.servicemesh.envoycontrol.services.MultiClusterState.Companion.toMultiClusterState
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceInstance
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceInstances
+import pl.allegro.tech.servicemesh.envoycontrol.services.ServiceName
 import pl.allegro.tech.servicemesh.envoycontrol.services.ServicesState
 import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.clusters.EnvoyClustersFactory
 import pl.allegro.tech.servicemesh.envoycontrol.snapshot.resource.endpoints.EnvoyEndpointsFactory
@@ -58,6 +59,7 @@ import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import reactor.core.scheduler.Schedulers
 import java.time.Duration
+import java.util.concurrent.ConcurrentHashMap
 import java.util.concurrent.CountDownLatch
 import java.util.concurrent.Executors
 import java.util.concurrent.TimeUnit
@@ -89,7 +91,7 @@ class SnapshotUpdaterTest {
 
     val clusterWithEnvoyInstances = ClusterState(
         ServicesState(
-            serviceNameToInstances = mapOf(
+            serviceNameToInstances = concurrentMapOf(
                 "service" to ServiceInstances(
                     "service", setOf(
                         ServiceInstance(
@@ -400,7 +402,7 @@ class SnapshotUpdaterTest {
 
         val clusterWithNoInstances = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service" to ServiceInstances("service", setOf())
                 )
             ),
@@ -443,7 +445,7 @@ class SnapshotUpdaterTest {
 
         val clusterWithOrder = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service" to ServiceInstances("service", setOf()),
                     "service2" to ServiceInstances("service2", setOf())
                 )
@@ -452,7 +454,7 @@ class SnapshotUpdaterTest {
         ).toMultiClusterState()
         val clusterWithoutOrder = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service2" to ServiceInstances("service2", setOf()),
                     "service" to ServiceInstances("service", setOf())
                 )
@@ -493,7 +495,7 @@ class SnapshotUpdaterTest {
 
         val clusterLocal = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service" to ServiceInstances(
                         "service", setOf(
                             ServiceInstance(
@@ -512,7 +514,7 @@ class SnapshotUpdaterTest {
 
         val remoteClusterWithBothServices = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service" to ServiceInstances("service", setOf()),
                     "servicePresentInJustOneRemote" to ServiceInstances("servicePresentInJustOneRemote", setOf())
                 )
@@ -522,7 +524,7 @@ class SnapshotUpdaterTest {
 
         val remoteClusterWithJustOneService = ClusterState(
             ServicesState(
-                serviceNameToInstances = mapOf(
+                serviceNameToInstances = concurrentMapOf(
                     "service" to ServiceInstances("service", setOf())
                 )
             ),
@@ -584,7 +586,7 @@ class SnapshotUpdaterTest {
         )
 
         val stateWithNoServices = ClusterState(
-            ServicesState(serviceNameToInstances = mapOf()),
+            ServicesState(serviceNameToInstances = concurrentMapOf()),
             Locality.LOCAL, "cluster"
         ).toMultiClusterState()
 
@@ -988,7 +990,7 @@ class SnapshotUpdaterTest {
     private fun fluxOfServices(vararg services: String) = Flux.just(createClusterState(*services).toMultiClusterState())
     private fun createClusterState(vararg services: String) = ClusterState(
         ServicesState(
-            serviceNameToInstances = services.map { it to ServiceInstances(it, emptySet()) }.toMap()
+            serviceNameToInstances = ConcurrentHashMap(services.associateWith { ServiceInstances(it, emptySet()) })
 
         ),
         Locality.LOCAL, "cluster")
@@ -1244,6 +1246,12 @@ class SnapshotUpdaterTest {
         snapshotChangeAuditor = snapshotChangeAuditor,
         globalSnapshotAuditScheduler = Schedulers.newSingle("audit-snapshot")
     )
+
+    private fun concurrentMapOf(vararg elements: Pair<ServiceName, ServiceInstances>): ConcurrentHashMap<ServiceName, ServiceInstances> {
+        val state = ConcurrentHashMap<ServiceName, ServiceInstances>()
+        elements.forEach { (name, instance) -> state[name] = instance }
+        return state
+    }
 }
 
 fun serviceDependencies(vararg dependencies: Pair<String, Outgoing.TimeoutPolicy?>): Set<ServiceDependency> =