+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Changed
-
 - Global snapshot auditor added
-- Changelog added
+- updated project java version 11
+
+## [0.19.5]
+### Changed 
+- Fixed issue with ConcurrentModificationException when using new ServicesState
+
+## [0.19.3]
+### Changed
+- Nothing. Empty commit for fix released version.